requirements.txt
astroid==2.2.5
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
gluoncv==0.5.0
graphviz==0.8.4
idna==2.8
isort==4.3.21
joblib==0.13.2
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
matplotlib==3.1.1
mccabe==0.6.1
mxnet-cu101==1.5.0
numpy==1.17.2
opencv-python==4.1.1.26
pandas==0.25.1
Pillow==6.1.0
pylint==2.3.1
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
tqdm==4.36.0
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
